Code Commit

+++ app/src/main/java/th/service/data/ThSense.java
@@ -42,7 +42,7 @@ public class ThSense {
         thSense.setType((byte) 100);
         thSense.setSubType((byte) 101);
         thSense.setExtType((byte)102);
-        thSense.setView((byte) value);
+        thSense.setView(value);
         thSense.setSense(ConvertUtils.intTo2Bytes(value));
         thSense.setSenseMax(ConvertUtils.intTo2Bytes(1000));
         thSense.setSenseMin(ConvertUtils.intTo2Bytes(0));
